requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
openai==0.28.1
lxml==4.9.3
python-dateutil==2.8.2
waitress==2.1.2